Add watsonAnalysis for speech to text
