+++ server/controllers/audio.controller.js
@@ -1,16 +1,19 @@
 var Promise = require('bluebird');
 var ffmpeg = require('fluent-ffmpeg');
 var path = require('path');
-var fs = require('fs');
-var watson = require('watson-developer-cloud');
-var apiKeys = '../config/localConfig.js';
+var watsonAnalysis = require('../config/localConfig.js');
 
-var speechToText = watson.speech_to_text(apiKeys.watsonCredentials);
 
-/*
 extractAudio(videoURL)
-  .then(watsonSpeechToText, function() { console.log('rejected') });
-  */
+  .then(function(wavFile){
+    beyondVerbal(wavFile);              //save beyondVerbalResults to db
+    watsonAnalysis.watsonSpeechToText(wavFile)         //save watsonResults to db
+      .then(function(watsonResults){
+        alchemyAnalysis(watsonResults); //save alchemyResults to db
+      })
+  });
+
+  
 
 module.exports.extractAudio = function(videoURL) {
   return new Promise(function(resolve, reject){
@@ -28,39 +31,7 @@ module.exports.extractAudio = function(videoURL) {
   })
 };
 
-module.exports.watsonSpeechToText = function(audioFile) {
-  return new Promise(function(resolve, reject){
-    var results = [];
-    var params = {
-      content_type: 'audio/wav',
-      timestamps: true,
-      continuous: true
-    };
-    
-    //create stream
-    var recognizeStream = speechToText.createRecognizeStream(params);
-    //pipe in audio
-    fs.createReadStream(audioFile).pipe(recognizeStream);
-    //saves text only to file
-    recognizeStream.pipe(fs.createWriteStream(__dirname + '/wavFiles/' + 'transcription.txt'));
-    // to get strings instead of Buffers from `data` events
-    recognizeStream.setEncoding('utf8');
-    recognizeStream.on('results', function(e){
-      if(e.results[0].final) {
-        results.push(e);
-      }
-    });
 
-    recognizeStream.on('error', function(err) {
-      console.log('Error writing to transcript.json: ' + err));
-    });
-
-    recognizeStream.on('connection-close', function() {
-      var verbalAnalysis = JSON.stringify(results);
-      resolve(verbalAnalysis);
-    });
-  });
-};
